pyproject.toml
[tool.poetry]
name = "haus-api"
version = "0.1.0"
description = "HAUS REST backend"
authors = ["iTecAI <matteovh@gmail.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
beanie = "^1.24.0"
motor = "^3.3.2"
litestar = {version = "^2.4.5", extras = ["cryptography", "standard"]}
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
